[#4876] Add support for modifier keys during drag operations
Adds the ability to use the OS-defined modifier key (usually
Alt on Windows and Option on Mac) to toggle between the default
drop behavior and the opposite behavior. So if dragging within
the same actor this changes from the default move behaior to
copy behavior and the opposite when dragging between different
actors or to the sidebar.

Enabling this required access to the current drag payload during
the `ondragover` event, so this extends the `DragDrop` class
provided by core to store that information during the `ondragstart`
event and adds a new handler for the `ondragend` event to clear
the stored payload.

Currently this only covers dragging items, with some minor
improvements to dragging favorites on the character sheet.
This framework could be expanded in the future to support
dragging actors into the bastion tab as well as dragging active
effects, advancements, or activities.

Closes #4876